LAB 08 - Caio Pantaleao
I started a session on the process.php file to be able to store the first name variable inside a session. This allowed me to call this variable in other files (view.php and add.php). I also created a destroy.php file that allow the user to finish (destroy) his session. A border was added to the table for better display.