LAB 08 - Caio Pantaleao
I started a session in the process.php file and stored the first name variable inside a session so I could display it on different files (view.php and add.php). I also created a destroy.php file to allow the user to destroy their session through the view.php file and added a boarder to the table so that the information could be better displayed. Once the session was destroyed the user would be redirected to the index.php page.